Combine PapyrosStateManager and StatusPanel into RunStateManager
